Broom methods for cureit objects

Broom methods for cureit objects

Usage

# S3 method for class 'cureit'
tidy(x, exponentiate = FALSE, conf.int = FALSE, conf.level = 0.95, ...)

Arguments

x

An smcure object created by smcure::smcure()

exponentiate

Logical indicating whether or not to exponentiate the coefficient estimates. Defaults to FALSE.

conf.int

A Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

Level of the confidence interval. Default matches that in smcure(conf.level=) (typically, 0.95)

...

Additional arguments passed to other methods.

Value

a tibble

See also

Examples

cureit <- cureit(surv_formula = Surv(ttdeath, death) ~ marker, cure_formula = ~marker,
 data = trial)
#> Warning: 0 of 100 did not converge.

tidy(cureit)
#> # A tibble: 3 × 5
#>   term                    estimate std.error statistic p.value
#>   <chr>                      <dbl>     <dbl>     <dbl>   <dbl>
#> 1 (Intercept), Cure model   0.293      0.167     1.76   0.0782
#> 2 marker, Cure model       -0.112      0.185    -0.608  0.543 
#> 3 marker, Survival model    0.0132     0.117     0.112  0.911