Predicted survival probability by cure model

Predicted survival probability by cure model

Usage

# S3 method for cureit
predict(
  object,
  times = NULL,
  probs = NULL,
  newdata = NULL,
  method = "prob",
  brier = FALSE,
  cox = FALSE,
  ...
)

Arguments

object

input object

times

Numeric vector of times to obtain survival probability estimates at

probs

Numeric vector of quantiles to obtain estimates at

newdata

A base::data.frame() or tibble::tibble() containing all the original predictors used to create x. Defaults to NULL.

method

Output format of predicted values: "lp" (linear predictor) or "prob" (predicted probabilities).

brier

Boolean indicator of calculating the Brier scores at specified times.

cox

Boolean indicator of fitting the Cox model for the training data and calculating the Brier scores at specified times for newdata.

...

passed to methods

Value

named list of prediction estimates

See also

Examples

p <- cureit(surv_formula = Surv(ttdeath, death) ~ age, 
   cure_formula = ~ age,
   data = trial) 
#> ✖ There was an error evaluating the LHS of the formula.
#> Warning: additional arguments ignored in stop()
#> Error in rlang::eval_tidy(., data = data): object 'trial' not found
   
predict(p, times = 12, newdata = trial[1:10, ])
#> Error in predict(p, times = 12, newdata = trial[1:10, ]): object 'p' not found